requirement.txt
asgiref==3.6.0
cffi==1.15.1
distlib==0.3.6
Django==4.1.7
django-braces==1.15.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.9.1
gunicorn==21.2.0
Markdown==3.5.1
packaging==23.2
Pillow==9.4.0
platformdirs==3.1.1
pycparser==2.21
pytz==2023.3.post1
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.21.0
whitenoise==6.4.0
